+++ changelog.md
@@ -1,7 +1,7 @@
 # Change Log
 
 #v0.9.1
-- Artisan command 'db:slow-log-queries'.
+- Artisan command 'db:log-slow-queries'.
 
 #v0.9
 - Minimum Laravel version: 8.